+++ Changelog.md
@@ -2,12 +2,20 @@
 
 
 ## TODO
-* Error handling if not signed-in
-* Omnibar: open in new tab search results
 * Open Advanced search
-
 * Context Menu: Insert Image by url
 
+## 2024-03-14
+* Fix: Remove Context menus in Extension Action menu.
+* Added context menus for Help and Release Notes to extension action menu
+* Change mouse pointer on clickable images
+* Focus on Next after search. Allows quick navigation with Tab to search results
+* Next and Prev arrow focusable with Tab and clickable with Enter
+* Added pagination and limit option
+* Error handling if not signed-in
+
+## 2024-03-13
+* * Omnibar: open in new tab search results
 ## 2024-03-12
 * First Release implementing Quick Search in popup
 * Context Menu: Number Headings